requirements.txt
Flask
gunicorn
requests
pymongo
dnspython 
python-dotenv